Update baseof.html
Made a small change (replaced '.' with '$') to allow the full page context to be passed to a custom sidebar partial from the following line:

{{- partial (printf "sidebar/%s.html" .) $ }} 

Previously, only a string was passed to the sidebar partial. The change I'm proposing will allow coders to create a sidebar partial that will list all the tags used in all blog posts. This wouldn't be possible if only a string was passed to the sidebar partial.